+++ core/src/main/java/com/picus/core/domain/chat/domain/entity/message/MessageImageResource.java
@@ -1,6 +1,6 @@
 package com.picus.core.domain.chat.domain.entity.message;
 
-import com.picus.core.global.common.image.domain.entity.Image;
+import com.picus.core.domain.shared.image.domain.entity.Image;
 import jakarta.persistence.Entity;
 import lombok.AccessLevel;
 import lombok.Getter;
