+++ core/src/main/java/com/picus/core/domain/post/application/usecase/ViewTrackerUseCase.java
@@ -1,12 +1,14 @@
 package com.picus.core.domain.post.application.usecase;
 
 import com.picus.core.domain.post.domain.entity.view.ViewCount;
-import com.picus.core.global.common.exception.PostNotFoundException;
+import com.picus.core.global.common.exception.RestApiException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
+
+import static com.picus.core.global.common.exception.code.status.GlobalErrorStatus._NOT_FOUND;
 
 @Service
 @RequiredArgsConstructor
@@ -16,7 +18,7 @@ public class ViewTrackerUseCase {
 
     public Integer findViewCount(Long postId) {
         return Optional.ofNullable(redisTemplate.opsForValue().get(ViewCount.generateKey(postId)))
-                .orElseThrow(PostNotFoundException::new);
+                .orElseThrow(() -> new RestApiException(_NOT_FOUND));
     }
 
     private void incrementViewCount(Long postId) {
