+++ core/src/main/java/com/picus/core/domain/post/domain/entity/category/PostCategory.java
@@ -2,7 +2,7 @@ package com.picus.core.domain.post.domain.entity.category;
 
 import com.picus.core.domain.post.domain.entity.Post;
 import com.picus.core.global.common.base.BaseEntity;
-import com.picus.core.global.common.category.entity.Category;
+import com.picus.core.domain.shared.category.entity.Category;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
