+++ core/src/main/java/com/picus/core/domain/shared/area/entity/Region.java
@@ -1,0 +1,26 @@
+package com.picus.core.domain.shared.area.entity;
+
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Getter
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Region {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "region_id")
+    private Long id;
+
+    private String name;
+
+    @OneToMany(cascade = {CascadeType.PERSIST, CascadeType.REMOVE},
+            orphanRemoval = true, fetch = FetchType.LAZY, mappedBy = "region")
+    private List<District> districts = new ArrayList<>();
+}
