+++ core/src/main/java/com/picus/core/domain/shared/category/entity/Type.java
@@ -1,0 +1,8 @@
+package com.picus.core.domain.shared.category.entity;
+
+public enum Type {
+
+    PLACE,  // 장소
+    THEME,  // 테마
+    MOOD    // 분위기
+}
