+++ core/src/main/java/com/picus/core/domain/shared/enums/ApprovalStatus.java
@@ -1,0 +1,7 @@
+package com.picus.core.domain.shared.enums;
+
+public enum ApprovalStatus {
+    PENDING,
+    APPROVAL,
+    REJECTED
+}
