+++ core/src/main/java/com/picus/core/domain/shared/image/application/dto/response/ImageUrl.java
@@ -1,0 +1,5 @@
+package com.picus.core.domain.shared.image.application.dto.response;
+
+public record ImageUrl (
+        String url
+) {}
