+++ core/src/main/java/com/picus/core/domain/shared/image/application/dto/response/UploadUrl.java
@@ -1,0 +1,9 @@
+package com.picus.core.domain.shared.image.application.dto.response;
+
+import lombok.Builder;
+
+@Builder
+public record UploadUrl (
+        String preSignedUrl,
+        Long imageId
+) {}
