+++ core/src/main/java/com/picus/core/domain/shared/image/domain/entity/ImageType.java
@@ -1,0 +1,9 @@
+package com.picus.core.domain.shared.image.domain.entity;
+
+public enum ImageType {
+    BACKGROUND,
+    PROFILE,
+    MESSAGE,
+    POST,
+    REVIEW
+}
