+++ core/src/main/java/com/picus/core/domain/shared/image/domain/factory/strategy/ImageCreationStrategy.java
@@ -1,0 +1,8 @@
+package com.picus.core.domain.shared.image.domain.factory.strategy;
+
+import com.picus.core.domain.shared.image.application.dto.request.UploadImage;
+import com.picus.core.domain.shared.image.domain.entity.Image;
+
+public interface ImageCreationStrategy {
+    Image toEntity(String key, UploadImage request);
+}
