+++ core/src/main/java/com/picus/core/domain/shared/image/domain/factory/strategy/PostImageCreationStrategy.java
@@ -1,0 +1,24 @@
+package com.picus.core.domain.shared.image.domain.factory.strategy;
+
+import com.picus.core.domain.post.domain.entity.image.PostImageResource;
+import com.picus.core.domain.shared.image.application.dto.request.UploadImage;
+import com.picus.core.domain.shared.image.domain.entity.Image;
+import lombok.AccessLevel;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor(access = AccessLevel.PRIVATE)
+public class PostImageCreationStrategy implements ImageCreationStrategy {
+
+    private static final PostImageCreationStrategy INSTANCE = new PostImageCreationStrategy();
+
+    public static PostImageCreationStrategy getInstance() {
+        return INSTANCE;
+    }
+
+    @Override
+    public Image toEntity(String key, UploadImage request) {
+        return new PostImageResource(key, request.postNo());
+    }
+}
