+++ core/src/main/java/com/picus/core/domain/shared/image/domain/repository/ImageRepository.java
@@ -1,0 +1,7 @@
+package com.picus.core.domain.shared.image.domain.repository;
+
+import com.picus.core.domain.shared.image.domain.entity.Image;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ImageRepository extends JpaRepository<Image, Long> {
+}
