+++ core/src/main/java/com/picus/core/domain/shared/image/domain/repository/ReviewImageResourceRepository.java
@@ -1,0 +1,7 @@
+package com.picus.core.domain.shared.image.domain.repository;
+
+import com.picus.core.domain.review.entity.ReviewImageResource;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ReviewImageResourceRepository extends JpaRepository<ReviewImageResource, Long> {
+}
