+++ core/src/main/java/com/picus/core/domain/studio/entity/Studio.java
@@ -1,7 +1,7 @@
 package com.picus.core.domain.studio.entity;
 
 import com.picus.core.global.common.base.BaseEntity;
-import com.picus.core.global.common.enums.ApprovalStatus;
+import com.picus.core.domain.shared.enums.ApprovalStatus;
 import io.hypersistence.utils.hibernate.id.Tsid;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
