--- core/src/main/java/com/picus/core/global/common/area/entity/District.java
@@ -1,23 +1,0 @@
-package com.picus.core.global.common.area.entity;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class District {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "district_id")
-    private Long id;
-
-    private String name;
-
-    @ManyToOne
-    @JoinColumn(name = "region_id", nullable = false)
-    private Region region;
-}
