--- core/src/main/java/com/picus/core/global/common/category/entity/Type.java
@@ -1,8 +1,0 @@
-package com.picus.core.global.common.category.entity;
-
-public enum Type {
-
-    PLACE,  // 장소
-    THEME,  // 테마
-    MOOD    // 분위기
-}
