--- core/src/main/java/com/picus/core/global/common/enums/ApprovalStatus.java
@@ -1,7 +1,0 @@
-package com.picus.core.global.common.enums;
-
-public enum ApprovalStatus {
-    PENDING,
-    APPROVAL,
-    REJECTED
-}
