--- core/src/main/java/com/picus/core/global/common/image/application/dto/response/ImageUrl.java
@@ -1,5 +1,0 @@
-package com.picus.core.global.common.image.application.dto.response;
-
-public record ImageUrl (
-        String url
-) {}
