--- core/src/main/java/com/picus/core/global/common/image/application/dto/response/UploadUrl.java
@@ -1,9 +1,0 @@
-package com.picus.core.global.common.image.application.dto.response;
-
-import lombok.Builder;
-
-@Builder
-public record UploadUrl (
-        String preSignedUrl,
-        Long imageId
-) {}
