--- core/src/main/java/com/picus/core/global/common/image/domain/entity/ImageType.java
@@ -1,9 +1,0 @@
-package com.picus.core.global.common.image.domain.entity;
-
-public enum ImageType {
-    BACKGROUND,
-    PROFILE,
-    MESSAGE,
-    POST,
-    REVIEW
-}
