--- core/src/main/java/com/picus/core/global/common/image/domain/repository/ReviewImageResourceRepository.java
@@ -1,7 +1,0 @@
-package com.picus.core.global.common.image.domain.repository;
-
-import com.picus.core.domain.review.entity.ReviewImageResource;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ReviewImageResourceRepository extends JpaRepository<ReviewImageResource, Long> {
-}
